Finished inital Home Page Frontend design
Set up inital Home page front end design that includes a view of all
your lists a view of all your tasks and a view of an individual task in
more detail. Additionally includes pop ups for adding tasks and lists.
The selection on the tasks temporarily breaks the delete button but I
think this is worth the sacrifice for the demo